Fix: Refactor Simulator API endpoints
Refactor of API endpoint handlers to include response codes from API specification

+++ itu-minitwit/src/MiniTwit.Web/Api.cs
@@ -1,7 +1,12 @@
+using Azure.Core;
 using Chirp.Infrastructure.Chirp.Repositories;
 using Chirp.Infrastructure.DataTransferObjects;
+using Microsoft.AspNetCore.Authorization.Infrastructure;
 using Microsoft.AspNetCore.Identity.Data;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Build.Experimental.ProjectCache;
+using SQLitePCL;
+
 
 namespace Chirp.Web;
 
@@ -19,34 +24,53 @@ public static class Api
     
     public static void MapProductEndpoints(this WebApplication app)
     {
+        /* Returns list of author followed by username*/
         app.MapGet("/fllws/{username}",
-            (string username,[FromQuery (Name = "latest")] int? latests,[FromQuery (Name = "no")] int? no, IFollowRepository followRepository) =>
+            (string username,[FromHeader (Name = "Authorization")] string authorization, [FromQuery (Name = "latest")] int? latests,[FromQuery (Name = "no")] int? no, IFollowRepository followRepository,  IAuthorRepository authorRepository) =>
             {
                 if (latests.HasValue)
                     Latest = latests.Value;
 
+                if (authorization != "Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh")
+                {
+                    return Results.StatusCode(403);
+                } 
+
+                var author = authorRepository.GetAuthorByName(username).Result;
+                if (author == null)
+                {
+                    return Results.StatusCode(404);
+                }
+
                 var res = followRepository.GetFollowed(username).Result.Select(follow => follow.Followed).ToList();
-                
-                return new GetFollowsRequest(res);
+                return Results.Ok(res);
             });
-            
+
+        /* User follows/unfollows author*/    
         app.MapPost("/fllws/{username}",
-            (string username, [FromQuery (Name = "latest")] int? latests ,[FromBody] FollowRequest request, IFollowRepository followRepository) =>
+            (string username, [FromHeader (Name = "Authorization")] string authorization, [FromQuery (Name = "latest")] int? latests ,[FromBody] FollowRequest request, IFollowRepository followRepository, IAuthorRepository authorRepository) =>
             {
                 if (latests.HasValue)
                     Latest = latests.Value;
-                
+
+                if (authorization != "Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh")
+                {
+                    return Results.StatusCode(403);
+                } 
+
                 if (!string.IsNullOrEmpty(request.Follow))
                 {
-                    return followRepository.AddFollowing(username,request.Follow);
+                    var res = followRepository.AddFollowing(username,request.Follow);
+                    return Results.NoContent();
                 }
 
                 if (!string.IsNullOrEmpty(request.Unfollow))
                 {
-                    return followRepository.RemoveFollowing(username,request.Unfollow);
+                    var res = followRepository.RemoveFollowing(username,request.Unfollow);
+                    return Results.NoContent();
                 }
                 
-                return Task.FromResult(Results.BadRequest("Must specify either 'Follow' or 'Unfollow'"));
+                return Results.BadRequest("Must specify either 'Follow' or 'Unfollow'");
             });
         
         app.MapGet("/latest",() => new {Latest});
@@ -60,12 +84,20 @@ public static class Api
             });
         
         app.MapGet("/msgs/{username}",
-            (string username,[FromQuery (Name = "latest")] int? latests,[FromQuery (Name = "no")] int? no, ICheepRepository cheepRepository) =>
+            (string username,[FromHeader (Name = "Authorization")] string authorization, [FromQuery (Name = "latest")] int? latests,[FromQuery (Name = "no")] int? no, ICheepRepository cheepRepository) =>
             {
                 if (latests.HasValue)
+                {
                     Latest = latests.Value;
+                }
+
+                if (authorization != "Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh")
+                {
+                    return Results.StatusCode(403);
+                }  
                 
-                return cheepRepository.GetAllCheepsFromAuthor(username).Result.Select(cheep => new GetMessagesRequest(cheep.Text, cheep.Author.Name));
+                var res = cheepRepository.GetAllCheepsFromAuthor(username).Result.Select(cheep => new GetMessagesRequest(cheep.Text, cheep.Author.Name));
+                return Results.Ok(res);
             });
         
         
@@ -80,13 +112,11 @@ public static class Api
                 if (authorization != "Basic c2ltdWxhdG9yOnN1cGVyX3NhZmUh")
                 {
                     return Results.StatusCode(403);
-            
                 }    
 
-
                 var author = authorRepository.GetAuthorByName(username).Result;
-                
                 cheepRepository.AddCheep(msgRequest.Content, author);
+
                 return Results.NoContent();
                 
             });
